Color components should really be u8s
Then we can use saturating_add and saturating_sub! Why isn't there
saturating mult and div and mod tho?!?!

+++ src/value_tokenizer.rs
@@ -88,7 +88,7 @@ impl<'a> ValueTokenizer<'a> {
         Ok(ret)
     }
 
-    fn eat_integer(&mut self) -> i32 {
+    fn eat_integer(&mut self) -> u8 {
         let limit = self.value_str.len();
         let mut i = self.offset;
         let color_start = i;
